Update profile() to use deepcopy of model
Reloading the state_dict of a profiled model fails when the new model instance was not profiled as well. Profiling alters the model state_dict by adding buffers. Profiling a model should not change the model in any way. Maybe there is a more elegant way to to this without deepcopying (remove hooks etc.)